perf(redis): create shared Redis client to avoid multiple connections
