+++ api/src/lib/redis.ts
@@ -1,0 +1,29 @@
+import Redis from 'ioredis';
+import { getRedisConnection } from '../utils/env';
+
+let redisClient: Redis | null = null;
+
+export function getRedisClient(): Redis {
+  if (!redisClient) {
+    const connection = getRedisConnection();
+    if ('url' in connection) {
+      redisClient = new Redis(connection.url, {
+        maxRetriesPerRequest: 3,
+        enableReadyCheck: true,
+        lazyConnect: true,
+      });
+    } else {
+      redisClient = new Redis({
+        host: connection.host,
+        port: connection.port,
+        username: connection.username,
+        password: connection.password,
+        maxRetriesPerRequest: 3,
+        enableReadyCheck: true,
+        lazyConnect: true,
+      });
+    }
+  }
+  return redisClient;
+}
+
